\name{CompositionalNMF-package}
\alias{CompositionalNMF-package}
\docType{package}
\title{
Nonnegative Matrix Factorization with Compositional Data
}

\description{
Simplicial nonnegative matrix factorization (SNMF) using many algorithms is implemented.
}

\details{
\tabular{ll}{
Package: \tab CompositionalNMF\cr
Type: \tab Package\cr
Version: \tab 1.0 \cr
Date: \tab 2026-08-05\cr
License: \tab GPL-2\cr
}
}

\author{
Michail Tsagris \email{mtsagris@uoc.gr}.
}

\section{Maintainers}{
Michail Tsagris <mtsagris@uoc.gr>
}

\references{
Alenazi A. and Tsagris M. (2026). Simplicial nonnegative matrix factorization with applications to high--dimensional and large--scale datasets. In preparation.

Wang H. Y., Yang Q., Qin H. and Zha H. (2008, July). Dirichlet component analysis: feature extraction for compositional data. In Proceedings of the 25th International Conference on Machine Learning (pp. 1128-1135).

Cutler A. and Breiman L. (1994). Archetypal analysis. Technometrics, 36(4): 338--347.
}

%\note{
%}


